+++ src/Sentry/Scope.cs
@@ -15,10 +15,9 @@ namespace Sentry
     /// Scope data is sent together with any event captured
     /// during the lifetime of the scope.
     /// </remarks>
-    public class Scope : IScope
+    public class Scope : IEventLike
     {
         internal SentryOptions Options { get; }
-        IScopeOptions IScope.ScopeOptions => Options;
 
         internal bool Locked { get; set; }
 
@@ -114,25 +113,40 @@ namespace Sentry
         public string? Environment { get; set; }
 
         /// <inheritdoc />
-        public SdkVersion Sdk { get; internal set; } = new();
+        public string? TransactionName { get; set; }
 
-        /// <inheritdoc />
-        public IEnumerable<string> Fingerprint { get; set; } = Enumerable.Empty<string>();
-
-        /// <inheritdoc />
-        public IEnumerable<Breadcrumb> Breadcrumbs { get; } = new ConcurrentQueue<Breadcrumb>();
-
-        /// <inheritdoc />
-        public IReadOnlyDictionary<string, object?> Extra { get; } = new ConcurrentDictionary<string, object?>();
-
-        /// <inheritdoc />
-        public IReadOnlyDictionary<string, string> Tags { get; } = new ConcurrentDictionary<string, string>();
-
-        /// <inheritdoc />
+        /// <summary>
+        /// Transaction.
+        /// </summary>
         public Transaction? Transaction { get; set; }
 
         /// <inheritdoc />
-        public string? TransactionName { get; set; }
+        public SdkVersion Sdk { get; } = new();
+
+        /// <inheritdoc />
+        public IReadOnlyList<string> Fingerprint { get; set; } = Array.Empty<string>();
+
+        private readonly ConcurrentQueue<Breadcrumb> _breadcrumbs = new();
+
+        /// <inheritdoc />
+        public IReadOnlyCollection<Breadcrumb> Breadcrumbs => _breadcrumbs;
+
+        private readonly ConcurrentDictionary<string, object?> _extra = new();
+
+        /// <inheritdoc />
+        public IReadOnlyDictionary<string, object?> Extra => _extra;
+
+        private readonly ConcurrentDictionary<string, string> _tags = new();
+
+        /// <inheritdoc />
+        public IReadOnlyDictionary<string, string> Tags => _tags;
+
+        private readonly ConcurrentBag<Attachment> _attachments = new();
+
+        /// <summary>
+        /// Attachments.
+        /// </summary>
+        public IReadOnlyCollection<Attachment> Attachments => _attachments;
 
         /// <summary>
         /// Creates a scope with the specified options.
@@ -148,13 +162,146 @@ namespace Sentry
         {
         }
 
+        /// <inheritdoc />
+        public void AddBreadcrumb(Breadcrumb breadcrumb)
+        {
+            if (Options.BeforeBreadcrumb is { } beforeBreadcrumb)
+            {
+                if (beforeBreadcrumb(breadcrumb) is { } processedBreadcrumb)
+                {
+                    breadcrumb = processedBreadcrumb;
+                }
+                else
+                {
+                    // Callback returned null, which means the breadcrumb should be dropped
+                    return;
+                }
+            }
+
+            var overflow = Breadcrumbs.Count - Options.MaxBreadcrumbs + 1;
+
+            if (overflow > 0)
+            {
+                _breadcrumbs.TryDequeue(out _);
+            }
+
+            _breadcrumbs.Enqueue(breadcrumb);
+        }
+
+        /// <inheritdoc />
+        public void SetExtra(string key, object? value) => _extra[key] = value;
+
+        /// <inheritdoc />
+        public void SetTag(string key, string value) => _tags[key] = value;
+
+        /// <summary>
+        /// Adds an attachment.
+        /// </summary>
+        public void AddAttachment(Attachment attachment) => _attachments.Add(attachment);
+
+        /// <summary>
+        /// Applies the data from this scope to another event-like object.
+        /// </summary>
+        /// <param name="other">The scope to copy data to.</param>
+        /// <remarks>
+        /// Applies the data of 'from' into 'to'.
+        /// If data in 'from' is null, 'to' is unmodified.
+        /// Conflicting keys are not overriden.
+        /// This is a shallow copy.
+        /// </remarks>
+        public void Apply(IEventLike other)
+        {
+            // Not to throw on code that ignores nullability warnings.
+            // ReSharper disable once ConditionIsAlwaysTrueOrFalse
+            if (other is null)
+            {
+                return;
+            }
+
+            // Fingerprint isn't combined. It's absolute.
+            // One set explicitly on target (i.e: event)
+            // takes precedence and is not overwritten
+            if (!other.Fingerprint.Any() && Fingerprint.Any())
+            {
+                other.Fingerprint = Fingerprint;
+            }
+
+            foreach (var breadcrumb in Breadcrumbs)
+            {
+                other.AddBreadcrumb(breadcrumb);
+            }
+
+            foreach (var (key, value) in Extra)
+            {
+                if (!other.Extra.ContainsKey(key))
+                {
+                    other.SetExtra(key, value);
+                }
+            }
+
+            foreach (var (key, value) in Tags)
+            {
+                if (!other.Tags.ContainsKey(key))
+                {
+                    other.SetTag(key, value);
+                }
+            }
+
+            Contexts.CopyTo(other.Contexts);
+            Request.CopyTo(other.Request);
+            User.CopyTo(other.User);
+
+            other.Environment ??= Environment;
+            other.TransactionName ??= TransactionName;
+            other.Level ??= Level;
+
+            if (Sdk.Name is not null && Sdk.Version is not null)
+            {
+                other.Sdk.Name = Sdk.Name;
+                other.Sdk.Version = Sdk.Version;
+            }
+
+            foreach (var package in Sdk.InternalPackages)
+            {
+                other.Sdk.AddPackage(package);
+            }
+        }
+
+        /// <summary>
+        /// Applies data from one scope to another.
+        /// </summary>
+        public void Apply(Scope other)
+        {
+            // Not to throw on code that ignores nullability warnings.
+            // ReSharper disable once ConditionIsAlwaysTrueOrFalse
+            if (other is null)
+            {
+                return;
+            }
+
+            Apply((IEventLike)other);
+
+            other.Transaction ??= Transaction;
+
+            foreach (var attachment in Attachments)
+            {
+                other.AddAttachment(attachment);
+            }
+        }
+
+        /// <summary>
+        /// Applies the state object into the scope.
+        /// </summary>
+        /// <param name="state">The state object to apply.</param>
+        public void Apply(object state) => Options.SentryScopeStateProcessor.Apply(this, state);
+
         /// <summary>
         /// Clones the current <see cref="Scope"/>.
         /// </summary>
         public Scope Clone()
         {
             var clone = new Scope(Options);
-            this.Apply(clone);
+            Apply(clone);
 
             foreach (var processor in EventProcessors)
             {
@@ -189,8 +336,12 @@ namespace Sentry
                 }
                 catch (Exception e)
                 {
-                    this.AddBreadcrumb("Failed invoking event handler: " + e,
-                        level: BreadcrumbLevel.Error);
+                    AddBreadcrumb(
+                        new Breadcrumb(
+                            message: "Failed invoking event handler: " + e,
+                            level: BreadcrumbLevel.Error
+                        )
+                    );
                 }
                 finally
                 {