+++ test/Sentry.Tests/Protocol/BaseScopeTests.cs
@@ -5,7 +5,7 @@ namespace Sentry.Tests.Protocol
 {
     public class BaseScopeTests
     {
-        private readonly IScope _sut = new Scope(new SentryOptions());
+        private readonly Scope _sut = new Scope(new SentryOptions());
 
         [Fact]
         public void Fingerprint_ByDefault_ReturnsEmptyEnumerable()