+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Make all fields on UserFeedback optional. (#660) @Tyrrrz
 - Align transaction names with Java. (#659) @Tyrrrz
 - Include assembly name in default release. (#682) @Tyrrrz
+- Add support for attachments. (#670) @Tyrrrz
 - Improve logging for relay errors. (#683) @Tyrrrz
 - Report sentry.dotnet.aspnet on the new Sentry.AspNet package. (#681) @Tyrrrz